+++ packages/react/src/components/Container/Container.tsx
@@ -1,10 +1,11 @@
 import styles from './Container.module.scss'
+
 import type { ContainerProps } from './Container.types';
 
 export default function Container({ children }: ContainerProps) {
   return (
     <main className={`${styles['container']} mx-auto`}>
-      {children}
+        {children}
     </main>
   )
 }