+++ packages/react/src/types/index.ts
@@ -1,1 +1,1 @@
-export type * from './common.types'
+export type * from './common.types';